Another minor algorithm improvement
Pre-shift src by (1-exp1) to the right instead of shifting sqr(Sqrt) to
the left and then shifting adj to the right on each NR iteration.

Algorithmic change is minor but the speed-up is decent, esp. with gcc - more
than 5%. For comparison, the previous major algorithm change impprove speed
by ~10%.