Refactoring and minor algorithm changes
1. de-cpp. Convert all templates into normal inline functions

2. Split NR_step() in two - parts that depend on parameter N and parts
that are independent of N. The 2nd part does not have to be inlined
by compiler

3. Long addition/substraction at the end of NR step, where a single
signed word added to multiword number: do carry/borrow handling with
branch-based loop only when there is a need. In overwhelming majirity
of cases there is no carry/borrow out after the first addition, so branch
is very predictable and thus faster than long dependency chain of
branchless variant.