asm variant - do another part of the same changes as in CPP
Long addition/subtraction at the end of NR step, where a single
signed word added to multiword number: carry/borrow handling for
all but the first word is started by conditional branch and executed
only when there is actual non-zero carry or borrow. In overwhelming majority
of cases there is no carry/borrow out after the first addition, so branch
is very predictable and thus faster than long dependency chain of
fully branchless variant. In case there is non-zero carry the rest
proceeds in branchless chain, so the algorithm is not *exactly* the same
as in CPP version, but the difference should not affect average speed.

The same idea applied to final rounding - only the first addition is unconditional.
The remaining additions are done only when needed, after conditional branch.
This small optimization is not yet added to CPP variant.